requirements.txt
alembic==1.16.4
catboost==1.2.8
category_encoders==2.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
cmaes==0.12.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cramjam==2.11.0
cryptography==45.0.5
pytest==8.3.3
pytest-cov==6.0.0
psutil==6.1.0
cycler==0.12.1
fastparquet==2024.11.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.2.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
narwhals==2.0.1
networkx==3.5
numba==0.61.2
numpy==2.2.6
optuna==4.4.0
optunahub==0.3.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
plotly==6.2.0
pyarrow==21.0.0
pycparser==2.22
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rtdl_num_embeddings==0.0.12
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.42
statsmodels==0.14.5
sympy==1.14.0
tabm==0.0.1
threadpoolctl==3.6.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xgboost==3.0.3
